separate status to build w/o kubeapiserver
