+++ pkg/status/status.go
@@ -3,12 +3,13 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2018 Datadog, Inc.
 
+// +build !kubeapiserver
+
 package status
 
 import (
 	"encoding/json"
 	"expvar"
-	"fmt"
 	"os"
 	"strconv"
 	"time"
@@ -18,7 +19,6 @@ import (
 	"github.com/DataDog/datadog-agent/pkg/logs"
 	"github.com/DataDog/datadog-agent/pkg/metadata/host"
 	"github.com/DataDog/datadog-agent/pkg/util"
-	"github.com/DataDog/datadog-agent/pkg/util/kubernetes/apiserver/leaderelection"
 	"github.com/DataDog/datadog-agent/pkg/version"
 	"github.com/DataDog/gohai/platform"
 	log "github.com/cihub/seelog"
@@ -105,31 +105,6 @@ func GetCheckStatus(c check.Check, cs *check.Stats) ([]byte, error) {
 	return []byte(st), nil
 }
 
-// GetDCAStatus grabs the status from expvar and puts it into a map
-func GetDCAStatus() (map[string]interface{}, error) {
-	stats := make(map[string]interface{})
-	stats, err := expvarStats(stats)
-	if err != nil {
-		log.Errorf("Error Getting ExpVar Stats: %v", err)
-	}
-	stats["config"] = getPartialConfig()
-
-	stats["version"] = version.AgentVersion
-	stats["pid"] = os.Getpid()
-	hostname, err := util.GetHostname()
-	if err != nil {
-		log.Errorf("Error grabbing hostname for status: %v", err)
-		stats["metadata"] = host.GetPayloadFromCache("unknown")
-	} else {
-		stats["metadata"] = host.GetPayloadFromCache(hostname)
-	}
-	now := time.Now()
-	stats["time"] = now.Format(timeFormat)
-	stats["leaderelection"] = getLeaderElectionDetails()
-
-	return stats, nil
-}
-
 // getPartialConfig returns config parameters of interest for the status page
 func getPartialConfig() map[string]string {
 	conf := make(map[string]string)
@@ -167,21 +142,4 @@ func expvarStats(stats map[string]interface{}) (map[string]interface{}, error) {
 	}
 
 	return stats, err
-}
-
-func getLeaderElectionDetails() map[string]string {
-	leaderElectionStats := make(map[string]string)
-
-	leaderElectionDetails, err := leaderelection.GetLeaderDetails()
-	if err != nil {
-		leaderElectionStats["status"] = "Failing"
-		leaderElectionStats["error"] = err.Error()
-		return leaderElectionStats
-	}
-	leaderElectionStats["leaderName"] = leaderElectionDetails.HolderIdentity
-	leaderElectionStats["acquiredTime"] = leaderElectionDetails.AcquireTime.Format(time.RFC1123)
-	leaderElectionStats["renewedTime"] = leaderElectionDetails.RenewTime.Format(time.RFC1123)
-	leaderElectionStats["transitions"] = fmt.Sprintf("%d transitions", leaderElectionDetails.LeaderTransitions)
-	leaderElectionStats["status"] = "Running"
-	return leaderElectionStats
 }
